resolve comment, make semantics behavior into XXXConfigurator
