+++ engine/src/flutter/shell/platform/android/BUILD.gn
@@ -382,10 +382,10 @@ android_java_sources = [
   "io/flutter/view/AccessibilityNodeConfigurator.java",
   "io/flutter/view/AccessibilityStringBuilder.java",
   "io/flutter/view/AccessibilityViewEmbedder.java",
+  "io/flutter/view/BaseRoleConfigurator.java",
   "io/flutter/view/ClassNameRoleConfigurator.java",
   "io/flutter/view/ComboBoxRoleConfigurator.java",
   "io/flutter/view/FlutterCallbackInformation.java",
-  "io/flutter/view/GenericRoleConfigurator.java",
   "io/flutter/view/MenuRoleConfigurator.java",
   "io/flutter/view/ProgressBarRoleConfigurator.java",
   "io/flutter/view/RoleConfiguratorFactory.java",
