+++ engine/src/flutter/shell/platform/android/io/flutter/view/ClassNameRoleConfigurator.java
@@ -7,7 +7,7 @@ package io.flutter.view;
 import android.view.accessibility.AccessibilityNodeInfo;
 
 /** Configurator that simply sets the class name of the accessibility node. */
-public class ClassNameRoleConfigurator implements AccessibilityNodeConfigurator {
+public class ClassNameRoleConfigurator extends BaseRoleConfigurator {
   private final String className;
 
   public ClassNameRoleConfigurator(String className) {
@@ -15,7 +15,7 @@ public class ClassNameRoleConfigurator implements AccessibilityNodeConfigurator 
   }
 
   @Override
-  public void configure(AccessibilityNodeInfo result, AccessibilityBridge.SemanticsNode node) {
+  protected void configureRole(AccessibilityNodeInfo result, AccessibilityBridge.SemanticsNode node) {
     result.setClassName(className);
   }
 }
