+++ engine/src/flutter/shell/platform/android/io/flutter/view/ComboBoxRoleConfigurator.java
@@ -10,9 +10,9 @@ import android.view.accessibility.AccessibilityNodeInfo;
  * Configurator for the {@link AccessibilityBridge.Role#COMBO_BOX} role. Sets the class name to
  * Spinner and indicates it can open a popup.
  */
-public class ComboBoxRoleConfigurator implements AccessibilityNodeConfigurator {
+public class ComboBoxRoleConfigurator extends BaseRoleConfigurator {
   @Override
-  public void configure(AccessibilityNodeInfo result, AccessibilityBridge.SemanticsNode node) {
+  protected void configureRole(AccessibilityNodeInfo result, AccessibilityBridge.SemanticsNode node) {
     result.setClassName("android.widget.Spinner");
     result.setCanOpenPopup(true);
   }
